+++ plugin/src/BeautifulMentionsPlugin.tsx
@@ -24,6 +24,7 @@ import React, {
 } from "react";
 import * as ReactDOM from "react-dom";
 import {
+  BeautifulMentionsItemData,
   BeautifulMentionsMenuItem,
   BeautifulMentionsPluginProps,
 } from "./BeautifulMentionsPluginProps";
@@ -32,7 +33,6 @@ import {
   $createBeautifulMentionNode,
   $isBeautifulMentionNode,
   BeautifulMentionNode,
-  MentionNodeDataValue,
 } from "./MentionNode";
 import { MenuOption, MenuTextMatch } from "./Menu";
 import { TypeaheadMenuPlugin } from "./TypeaheadMenuPlugin";
@@ -70,7 +70,7 @@ class MentionOption extends MenuOption {
     public readonly trigger: string,
     value: string,
     displayValue: string,
-    data?: { [key: string]: MentionNodeDataValue },
+    data?: { [key: string]: BeautifulMentionsItemData },
   ) {
     super(value, displayValue, data);
     this.menuItem = {