+++ plugin/src/BeautifulMentionsPluginProps.ts
@@ -1,5 +1,4 @@
 import { ComponentPropsWithRef, ElementType } from "react";
-import { MentionNodeDataValue } from "./MentionNode";
 
 /**
  * Represents a menu item for a mention.
@@ -21,7 +20,7 @@ export interface BeautifulMentionsMenuItem {
   /**
    * Additional data belonging to the mention.
    */
-  data?: { [key: string]: MentionNodeDataValue };
+  data?: { [key: string]: BeautifulMentionsItemData };
 }
 
 /**
@@ -45,8 +44,10 @@ export interface BeautifulMentionsComboboxItem {
   /**
    * Additional data belonging to the mention.
    */
-  data?: { [key: string]: MentionNodeDataValue };
+  data?: { [key: string]: BeautifulMentionsItemData };
 }
+
+export type BeautifulMentionsItemData = string | boolean | number | null;
 
 /**
  * The mention without the trigger. For example: "John". Either a string or
@@ -57,7 +58,7 @@ export type BeautifulMentionsItem =
   | string
   | {
       value: string;
-      [key: string]: MentionNodeDataValue;
+      [key: string]: BeautifulMentionsItemData;
     };
 
 /**
@@ -336,7 +337,7 @@ export type BeautifulMentionsPluginProps =
  * a mention in the editor.
  */
 export interface BeautifulMentionComponentProps<
-  T extends { [p: string]: MentionNodeDataValue } = {},
+  T extends { [p: string]: BeautifulMentionsItemData } = {},
 > extends Omit<ComponentPropsWithRef<any>, "value" | "data"> {
   /**
    * The trigger of the mention.