+++ plugin/src/ComboboxPlugin.tsx
@@ -23,6 +23,7 @@ import { useCallback, useEffect, useMemo, useState } from "react";
 import * as ReactDOM from "react-dom";
 import {
   BeautifulMentionsComboboxItem,
+  BeautifulMentionsItemData,
   BeautifulMentionsPluginProps,
 } from "./BeautifulMentionsPluginProps";
 import {
@@ -34,7 +35,6 @@ import {
 import { CAN_USE_DOM, IS_MOBILE } from "./environment";
 import { $insertTriggerAtSelection } from "./mention-commands";
 import { useIsFocused } from "./useIsFocused";
-import { MentionNodeDataValue } from "./MentionNode";
 
 interface ComboboxPluginProps
   extends Pick<
@@ -72,7 +72,7 @@ class ComboboxOption extends MenuOption {
     public readonly itemType: "trigger" | "value" | "additional",
     value: string,
     displayValue: string,
-    data: { [key: string]: MentionNodeDataValue } = {},
+    data: { [key: string]: BeautifulMentionsItemData } = {},
   ) {
     super(value, displayValue, data);
     this.comboboxItem = {