+++ plugin/src/MentionComponent.tsx
@@ -23,8 +23,11 @@ import {
   SELECTION_CHANGE_COMMAND,
 } from "lexical";
 import React, { ElementType, useCallback, useMemo, useState } from "react";
-import { BeautifulMentionComponentProps as CustomBeautifulMentionComponentProps } from "./BeautifulMentionsPluginProps";
-import { $isBeautifulMentionNode, MentionNodeDataValue } from "./MentionNode";
+import {
+  BeautifulMentionsItemData,
+  BeautifulMentionComponentProps as CustomBeautifulMentionComponentProps,
+} from "./BeautifulMentionsPluginProps";
+import { $isBeautifulMentionNode } from "./MentionNode";
 import { IS_IOS } from "./environment";
 import { getNextSibling, getPreviousSibling } from "./mention-utils";
 import { BeautifulMentionsThemeValues } from "./theme";
@@ -33,7 +36,7 @@ interface BeautifulMentionComponentProps {
   nodeKey: NodeKey;
   trigger: string;
   value: string;
-  data?: { [p: string]: MentionNodeDataValue };
+  data?: { [p: string]: BeautifulMentionsItemData };
   component?: ElementType<CustomBeautifulMentionComponentProps> | null;
   className?: string;
   classNameFocused?: string;