+++ plugin/src/MentionNode.tsx
@@ -9,17 +9,18 @@ import {
   type NodeKey,
 } from "lexical";
 import React, { ElementType } from "react";
-import { BeautifulMentionComponentProps } from "./BeautifulMentionsPluginProps";
+import {
+  BeautifulMentionComponentProps,
+  BeautifulMentionsItemData,
+} from "./BeautifulMentionsPluginProps";
 import MentionComponent from "./MentionComponent";
 import { BeautifulMentionsTheme } from "./theme";
-
-export type MentionNodeDataValue = string | boolean | number | null;
 
 export type SerializedBeautifulMentionNode = Spread<
   {
     trigger: string;
     value: string;
-    data?: { [p: string]: MentionNodeDataValue };
+    data?: { [p: string]: BeautifulMentionsItemData };
   },
   SerializedLexicalNode
 >;
@@ -29,7 +30,7 @@ function convertElement(domNode: HTMLElement): DOMConversionOutput | null {
     "data-lexical-beautiful-mention-trigger",
   );
   const value = domNode.getAttribute("data-lexical-beautiful-mention-value");
-  let data: { [p: string]: MentionNodeDataValue } | undefined = undefined;
+  let data: { [p: string]: BeautifulMentionsItemData } | undefined = undefined;
   const dataStr = domNode.getAttribute("data-lexical-beautiful-mention-data");
   if (dataStr) {
     try {
@@ -54,7 +55,7 @@ function convertElement(domNode: HTMLElement): DOMConversionOutput | null {
 export class BeautifulMentionNode extends DecoratorNode<React.JSX.Element> {
   __trigger: string;
   __value: string;
-  __data?: { [p: string]: MentionNodeDataValue };
+  __data?: { [p: string]: BeautifulMentionsItemData };
 
   static getType() {
     return "beautifulMention";
@@ -112,7 +113,7 @@ export class BeautifulMentionNode extends DecoratorNode<React.JSX.Element> {
   constructor(
     trigger: string,
     value: string,
-    data?: { [p: string]: MentionNodeDataValue },
+    data?: { [p: string]: BeautifulMentionsItemData },
     key?: NodeKey,
   ) {
     super(key);
@@ -159,12 +160,12 @@ export class BeautifulMentionNode extends DecoratorNode<React.JSX.Element> {
     self.__value = value;
   }
 
-  getData(): { [p: string]: MentionNodeDataValue } | undefined {
+  getData(): { [p: string]: BeautifulMentionsItemData } | undefined {
     const self = this.getLatest();
     return self.__data;
   }
 
-  setData(data?: { [p: string]: MentionNodeDataValue }) {
+  setData(data?: { [p: string]: BeautifulMentionsItemData }) {
     const self = this.getWritable();
     self.__data = data;
   }
@@ -204,7 +205,7 @@ export class BeautifulMentionNode extends DecoratorNode<React.JSX.Element> {
 export function $createBeautifulMentionNode(
   trigger: string,
   value: string,
-  data?: { [p: string]: MentionNodeDataValue },
+  data?: { [p: string]: BeautifulMentionsItemData },
 ): BeautifulMentionNode {
   const mentionNode = new BeautifulMentionNode(trigger, value, data);
   return $applyNodeReplacement(mentionNode);