+++ plugin/src/Menu.tsx
@@ -24,8 +24,8 @@ import {
   useRef,
   useState,
 } from "react";
+import { BeautifulMentionsItemData } from "./BeautifulMentionsPluginProps";
 import { CAN_USE_DOM, IS_MOBILE } from "./environment";
-import { MentionNodeDataValue } from "./MentionNode";
 
 const useLayoutEffectImpl: typeof useLayoutEffect = CAN_USE_DOM
   ? useLayoutEffect
@@ -74,7 +74,7 @@ export class MenuOption {
     /**
      * Additional data belonging to the option. For example: `{ id: 1 }`.
      */
-    public readonly data?: { [key: string]: MentionNodeDataValue },
+    public readonly data?: { [key: string]: BeautifulMentionsItemData },
   ) {
     this.key = !data ? value : JSON.stringify({ ...data, value });
     this.displayValue = displayValue ?? value;