+++ plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # lexical-beautiful-mentions
 
+## 0.1.21
+
+### Patch Changes
+
+- f60adca: fix(createBeautifulMentionNode): prevent re-creating the BeautifulMentionNode class on every call
+
+## 0.1.20
+
+### Patch Changes
+
+- 662e895: feat(): add null type to BeautifulMentionsItem data fields
+
 ## 0.1.19
 
 ### Patch Changes